apps/cost-fetcher/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
boto3==1.34.69
clickhouse-connect==0.7.0
python-dotenv==1.0.1
pydantic==2.6.4 